Add tests for partials, and add Berton lateral attenuation accounting method

+++ pyNA/src/noise_src_jl/smooth_max.jl
@@ -1,4 +1,4 @@
-function smooth_max(k_smooth::Float64, level_sideline)
+function smooth_max(k_smooth, level_sideline)
 
     # Compute lateral noise
     smooth_max = maximum(level_sideline) + 1/k_smooth.*log(sum(exp.(k_smooth*(level_sideline.-maximum(level_sideline)))))
